moved api func into api folder , hard coded location , commented out flat list, simplified code
